ml-service/requirements.txt
flask
flask-cors
openai-whisper
librosa
deepface
numpy
scipy
opencv-python